Break library into modules
