+++ CancerGov/_src/Scripts/NCI/Modules/floatingDelighter/index.scss
@@ -128,65 +128,6 @@
 
 }
 
-.floating-delighter--cts {
-
-  .floating-delighter__link {
-    background-color: #82368c;
-
-    &:hover, &:focus, &:active {
-      background-color: #3f0557;
-      text-decoration: none;
-    }
-  }
-  
-  .floating-delighter__icon {
-    
-    &::after {
-      content: '';
-      @include svg-sprite(search-large);
-      font-size: 1.1em;
-      
-      @include bp(small) {
-        font-size: .7em;
-      }
-    }
-  }
-
-  .floating-delighter__label {
-
-  }
-}
-
-.floating-delighter--social-media {
-  .floating-delighter__link {
-    background-color: #0D95A1;
-    font-size: .9em;
-
-    &:hover, &:focus, &:active {
-      background-color:#0b8590;
-      text-decoration: none;
-    }
-  }
-
-  .floating-delighter__icon {
-    
-    &::after {
-      content: '';
-      @include svg-sprite(social-calendar);
-      
-      @include bp(small) {
-        font-size: .7em;
-      }
-    }
-  }
-
-  .floating-delighter__label {
-    @include bp(xtra-small) {
-      font-size: 14px;
-    }
-  }
-}
-
 // 10/31/2017 hotfix to hide the feedback delighter upon launch of new survey tool
 .results-delighter-rail,
 .view-delighter-rail {
