+++ CancerGov/_src/Scripts/NCI/Modules/floatingDelighter/types/cts/index.js
@@ -1,0 +1,11 @@
+import './index.scss';
+
+export const cts = {
+    href: '/about-cancer/treatment/clinical-trials/search',
+    innerHTML: `
+        <div class="floating-delighter__icon">
+        </div>
+        <div class="floating-delighter__label">Find a Clinical Trial</div>
+    `,
+    classList: ['floating-delighter--cts']
+};
