+++ CancerGov/_src/Scripts/NCI/Modules/floatingDelighter/types/socialMedia/index.js
@@ -1,0 +1,11 @@
+import './index.scss';
+
+export const socialMedia = {
+    href: '/news-events/events/social-media',
+    innerHTML: `
+        <div class="floating-delighter__icon">
+        </div>
+        <div class="floating-delighter__label">NCI Social Media Events</div>
+    `,
+    classList: ['floating-delighter--social-media']
+};
