+++ CancerGov/_src/Scripts/NCI/Modules/floatingDelighter/types/socialMedia/index.scss
@@ -1,0 +1,33 @@
+@import '../../../../../../StyleSheets/variables';
+@import '../../../../../../StyleSheets/breakpoints';
+@import '../../../../../../StyleSheets/sprites/svg-sprite';
+
+.floating-delighter--social-media {
+    .floating-delighter__link {
+      background-color: #0D95A1;
+      font-size: .9em;
+  
+      &:hover, &:focus, &:active {
+        background-color:#0b8590;
+        text-decoration: none;
+      }
+    }
+  
+    .floating-delighter__icon {
+      
+      &::after {
+        content: '';
+        @include svg-sprite(social-calendar);
+        
+        @include bp(small) {
+          font-size: .7em;
+        }
+      }
+    }
+  
+    .floating-delighter__label {
+      @include bp(xtra-small) {
+        font-size: 14px;
+      }
+    }
+  }
